lab3-6-mikolaj-olesinski/requirements.txt
dvc
dvc-s3
pandas
scikit-learn
polars
pyyaml
jupyter
matplotlib
seaborn
nbconvert
spacy
contractions
tqdm
papermill
mlflow
gensim
ruff
shap